Move asyncInfer to a separated file
